+++ changelog.md
@@ -55,7 +55,7 @@
 
 - Added `os.isRelativeTo` to tell whether a path is relative to another
 
-- Added a `sugar.chainOn` macro for easy function chaining that's available
+- Added `sugar.chainOn` and `sugar.chainEval` for easy function chaining that's available
   everywhere, there is no need to concern your APIs with returning the first argument
   to enable "chaining" (as in the builder pattern), instead use this.